Remove references to SDK to avoid confusion
We are releasing a more comprehensive starter kit SDK which includes a full exchange backend implementing the Hydro API, and sample frontend which will communicate with it. This project is still useful for building your own client from scratch that is meant to talk to a server running the Hydro API, however to avoid confusion we will call this a javascript client instead of a full SDK going forward.